Moved useFocus into hooks folder
